+++ sismic/code/python.py
@@ -1,65 +1,21 @@
-import collections
-import copy
 from functools import partial
 from types import CodeType
-from typing import Any, Callable, Dict, Iterator, List, Mapping, Optional
+from typing import Any, Dict, List, Optional, Mapping, Iterator
 
 from . import Evaluator
+from .context import FrozenContext, EventContextProvider, TimeContextProvider
 from ..exceptions import CodeEvaluationError
-from ..model import (Event, InternalEvent, MetaEvent, Statechart, StateMixin, Transition)
+from ..model import Event, Statechart, Transition
 
 __all__ = ['PythonEvaluator']
-
-
-class FrozenContext(collections.Mapping):
-    """
-    A shallow copy of a context. The keys of the underlying context are
-    exposed as attributes.
-    """
-    __slots__ = ['__frozencontext']
-
-    def __init__(self, context: Mapping) -> None:
-        self.__frozencontext = {k: copy.copy(v) for k, v in context.items()}
-
-    def __getattr__(self, item):
-        try:
-            return self.__frozencontext[item]
-        except KeyError:
-            raise AttributeError('{} has no attribute {}'.format(self, item))
-
-    def __getstate__(self):
-        return self.__frozencontext
-
-    def __setstate__(self, state):
-        self.__frozencontext = state
-
-    def __getitem__(self, key):
-        return self.__frozencontext[key]
-
-    def __len__(self):
-        return len(self.__frozencontext)
-
-    def __iter__(self):
-        return iter(self.__frozencontext)
-
-
-def _create_send_function(event_list: List[Event]) -> Callable[..., None]:
-    def send(name, **kwargs):
-        event_list.append(InternalEvent(name, **kwargs))
-    return send
-
-def _create_notify_function(event_list: List[Event]) -> Callable[..., None]:
-    def notify(name, **kwargs):
-        event_list.append(MetaEvent(name, **kwargs))
-    return notify
 
 
 class PythonEvaluator(Evaluator):
     """
     A code evaluator that understands Python.
 
-    Depending on the method that is called, the context can expose additional values:
-
+    This evaluator exposes some additional functions/variables:
+    
     - On both code execution and code evaluation:
         - A *time: float* value that represents the current time exposed by interpreter clock.
         - An *active(name: str) -> bool* Boolean function that takes a state name and return *True* if and only
@@ -75,7 +31,7 @@ class PythonEvaluator(Evaluator):
         - A *setdefault(name:str, value: Any) -> Any* function that defines and returns variable *name* in
           the global scope if it is not yet defined.
     - On guard or contract evaluation:
-        - If the code is related to a transition, the *event: Event* that fires the transition is exposed.
+        - If the code is related to a transition, the *event: Event* that fires the transition is exposed.          
     - On guard or contract (except preconditions) evaluation:
         - An *after(sec: float) -> bool* Boolean function that returns *True* if and only if the source state
           was entered more than *sec* seconds ago. The time is evaluated according to Interpreter's clock.
@@ -98,7 +54,6 @@ class PythonEvaluator(Evaluator):
         is expected to be an *Interpreter* instance
     :param initial_context: a dictionary that will be used as *__locals__*
     """
-
     def __init__(self, interpreter=None, *, initial_context: Mapping[str, Any]=None) -> None:
         super().__init__(interpreter, initial_context=initial_context)
 
@@ -106,33 +61,23 @@ class PythonEvaluator(Evaluator):
         self._context.update(initial_context if initial_context else {})
         self._interpreter = interpreter
 
-        # Memory and entry time
-        self._memory = {}  # type: Dict[int, Mapping]
-        self._entry_time = {}  # type: Dict[str, float]
-        self._idle_time = {}  # type: Dict[str, float]
+        # Context providers
+        self._event_provider = EventContextProvider()
+        self._time_provider = TimeContextProvider()
+
+        self._interpreter.bind(self._event_provider, internal=False, meta=True)
+        self._interpreter.bind(self._time_provider, internal=False, meta=True)
 
         # Precompiled code
         self._evaluable_code = {}  # type: Dict[str, CodeType]
         self._executable_code = {}  # type: Dict[str, CodeType]
 
-        # Intercept sent and received events
-        self._sent_events = []  # type: List[Event]
-        if self._interpreter is not None:
-            self._interpreter.bind(self._sent_events.append)
-        self._received_event = None  # type: Optional[Event]
+        # Frozen context for __old__
+        self._memory = {}  # type: Dict[int, FrozenContext]
 
     @property
     def context(self) -> Mapping:
         return self._context
-
-    def on_step_starts(self, event: Optional[Event]=None) -> None:
-        """
-        Called each time the interpreter starts a macro step.
-
-        :param event: Optional processed event
-        """
-        self._sent_events.clear()
-        self._received_event = event
 
     def _setdefault(self, name: str, value: Any) -> Any:
         """
@@ -143,49 +88,6 @@ class PythonEvaluator(Evaluator):
         :return: value of the variable
         """
         return self._context.setdefault(name, value)
-
-    def _received(self, name: str) -> bool:
-        """
-        :param name: name of an event
-        :return: True if given event name was received in current step.
-        """
-        return getattr(self._received_event, 'name', None) == name
-
-    def _sent(self, name: str) -> bool:
-        """
-        :param name: name of an event
-        :return: True if given event name was sent during this step.
-        """
-        return any((name == e.name for e in self._sent_events))
-
-    def _active(self, name: str) -> bool:
-        """
-        Return True if given state name is active.
-
-        :param name: name of a state
-        :return: True if given state name is active.
-        """
-        return name in self._interpreter.configuration
-
-    def _after(self, name: str, seconds: float) -> bool:
-        """
-        Return True if given state was entered more than *seconds* ago.
-
-        :param name: name of a state
-        :param seconds: elapsed time
-        :return: True if given state was entered more than *seconds* ago.
-        """
-        return self._interpreter.time - seconds >= self._entry_time[name]
-
-    def _idle(self, name: str, seconds: float) -> bool:
-        """
-        Return True if given state was the target of a transition more than *seconds* ago.
-
-        :param name: name of a state
-        :param seconds: elapsed time
-        :return: True if given state was the target of a transition more than *seconds* ago.
-        """
-        return self._interpreter.time - seconds >= self._idle_time[name]
 
     def _evaluate_code(self, code: Optional[str], *, additional_context: Mapping[str, Any]=None) -> bool:
         """
@@ -203,8 +105,8 @@ class PythonEvaluator(Evaluator):
             compiled_code = self._evaluable_code.setdefault(code, compile(code, '<string>', 'eval'))
 
         exposed_context = {
-            'active': self._active,
-            'time': self._interpreter.time,
+            'active': self._time_provider.active,
+            'time': self._time_provider.time,
         }
         exposed_context.update(additional_context if additional_context is not None else {})
 
@@ -212,6 +114,7 @@ class PythonEvaluator(Evaluator):
             return eval(compiled_code, exposed_context, self._context)  # type: ignore
         except Exception as e:
             raise CodeEvaluationError('"{}" occurred while evaluating "{}"'.format(e, code)) from e
+
 
     def _execute_code(self, code: Optional[str], *, additional_context: Mapping[str, Any]=None) -> List[Event]:
         """
@@ -228,35 +131,20 @@ class PythonEvaluator(Evaluator):
         if compiled_code is None:
             compiled_code = self._executable_code.setdefault(code, compile(code, '<string>', 'exec'))
 
-        # Handle sent events
-        sent_events = []  # type: List[Event]
-
         exposed_context = {
-            'active': self._active,
-            'send': _create_send_function(sent_events),
-            'notify': _create_notify_function(sent_events),
-            'time': self._interpreter.time,
+            'active': self._time_provider.active,
+            'time': self._time_provider.time,
+            'send': self._event_provider.send,
+            'notify': self._event_provider.notify,
             'setdefault': self._setdefault,
         }
         exposed_context.update(additional_context if additional_context is not None else {})
 
         try:
             exec(compiled_code, exposed_context, self._context)  # type: ignore
-            return sent_events
+            return self._event_provider.pending
         except Exception as e:
             raise CodeEvaluationError('"{}" occurred while executing "{}"'.format(e, code)) from e
-
-    def execute_statechart(self, statechart: Statechart):
-        """
-        Execute the initial code of a statechart.
-        This method is called at the very beginning of the execution.
-
-        :param statechart: statechart to consider
-        """
-        if statechart.preamble:
-            events = self._execute_code(statechart.preamble)
-            if len(events) > 0:
-                raise CodeEvaluationError('Events cannot be raised by statechart preamble')
 
     def evaluate_guard(self, transition: Transition, event: Optional[Event]=None) -> bool:
         """
@@ -267,52 +155,11 @@ class PythonEvaluator(Evaluator):
         :return: truth value of *code*
         """
         additional_context = {
+            'after': partial(self._time_provider.after, transition.source),
+            'idle': partial(self._time_provider.idle, transition.source),
             'event': event,
-            'after': partial(self._after, transition.source),
-            'idle': partial(self._idle, transition.source),
         }
         return self._evaluate_code(getattr(transition, 'guard', None), additional_context=additional_context)
-
-    def execute_action(self, transition: Transition, event: Optional[Event]=None) -> List[Event]:
-        """
-        Execute the action for given transition.
-        This method is called for every transition that is processed, even those with no *action*.
-
-        :param transition: the considered transition
-        :param event: instance of *Event* if any
-        :return: a list of sent events
-        """
-        execution = self._execute_code(getattr(transition, 'action', None), additional_context={'event': event})
-
-        self._idle_time[transition.source] = self._interpreter.time
-
-        return execution
-
-    def execute_on_entry(self, state: StateMixin) -> List[Event]:
-        """
-        Execute the on entry action for given state.
-        This method is called for every state that is entered, even those with no *on_entry*.
-
-        :param state: the considered state
-        :return: a list of sent events
-        """
-        execution = self._execute_code(getattr(state, 'on_entry', None))
-
-        self._entry_time[state.name] = self._interpreter.time
-        self._idle_time[state.name] = self._interpreter.time
-
-        return execution
-
-
-    def execute_on_exit(self, state: StateMixin) -> List[Event]:
-        """
-        Execute the on exit action for given state.
-        This method is called for every state that is exited, even those with no *on_exit*.
-
-        :param state: the considered state
-        :return: a list of sent events
-        """
-        return self._execute_code(getattr(state, 'on_exit', None))
 
     def evaluate_preconditions(self, obj, event: Optional[Event]=None) -> Iterator[str]:
         """
@@ -323,13 +170,13 @@ class PythonEvaluator(Evaluator):
         :param event: an optional *Event* instance, in the case of a transition
         :return: list of unsatisfied conditions
         """
-        additional_context = {'event': event} if isinstance(obj, Transition) else {}  # type: Dict[str, Any]
-        additional_context.update({
-            'received': self._received,
-            'sent': self._sent
-        })  # type: ignore
+        additional_context = {
+            'received': self._event_provider.received,
+            'sent': self._event_provider.sent,
+            'event': event if isinstance(obj, Transition) else None,
+        }
 
-        # Only needed if there is an invariant, a postcondition or a sequential condition
+        # Deal with __old__ in contracts, only required if there is an invariant or a postcondition
         if len(getattr(obj, 'invariants', [])) > 0 or len(getattr(obj, 'postconditions', [])) > 0:
             self._memory[id(obj)] = FrozenContext(self._context)
 
@@ -349,14 +196,14 @@ class PythonEvaluator(Evaluator):
         """
         state_name = obj.source if isinstance(obj, Transition) else obj.name
 
-        additional_context = {'event': event} if isinstance(obj, Transition) else {}  # type: Dict[str, Any]
-        additional_context.update({
+        additional_context = {
             '__old__': self._memory.get(id(obj), None),
-            'after': partial(self._after, state_name),
-            'idle': partial(self._idle, state_name),
-            'received': self._received,
-            'sent': self._sent,
-        })
+            'after': partial(self._time_provider.after, state_name),
+            'idle': partial(self._time_provider.idle, state_name),
+            'received': self._event_provider.received,
+            'sent': self._event_provider.sent,
+            'event': event if isinstance(obj, Transition) else None,
+        }
 
         return filter(
             lambda c: not self._evaluate_code(c, additional_context=additional_context),
@@ -373,15 +220,15 @@ class PythonEvaluator(Evaluator):
         :return: list of unsatisfied conditions
         """
         state_name = obj.source if isinstance(obj, Transition) else obj.name
-
-        additional_context = {'event': event} if isinstance(obj, Transition) else {}  # type: Dict[str, Any]
-        additional_context.update({
+        
+        additional_context = {
             '__old__': self._memory.get(id(obj), None),
-            'after': partial(self._after, state_name),
-            'idle': partial(self._idle, state_name),
-            'received': self._received,
-            'sent': self._sent,
-        })
+            'after': partial(self._time_provider.after, state_name),
+            'idle': partial(self._time_provider.idle, state_name),
+            'received': self._event_provider.received,
+            'sent': self._event_provider.sent,
+            'event': event if isinstance(obj, Transition) else None,
+        }
 
         return filter(
             lambda c: not self._evaluate_code(c, additional_context=additional_context),